requirements.txt
pip3 freeze > requirements.txtbokeh==3.2.1
click==8.1.6
cloudpickle==2.2.1
contourpy==1.1.0
dask==2023.7.1
distributed==2023.7.1
fsspec==2023.6.0
importlib-metadata==6.8.0
Jinja2==3.1.2
locket==1.0.0
lz4==4.3.2
MarkupSafe==2.1.3
msgpack==1.0.5
numpy==1.25.1
packaging==23.1
pandas==2.0.3
partd==1.4.0
Pillow==10.0.0
postgres==4.0
psutil==5.9.5
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pyarrow==12.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sortedcontainers==2.4.0
tblib==2.0.0
toolz==0.12.0
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.4
xyzservices==2023.7.0
zict==3.0.0
zipp==3.16.2
